fix: Define PostHog.init at gem load time
PostHog.init was defined inside a Rails
initializer (`posthog.set_configs`), so it wasn't available until that
initializer ran. In some Rails boot sequences,
`config/initializers/posthog.rb` loaded first, causing `NoMethodError:
undefined method 'init' for module PostHog`.

This moves PostHog.init and delegated methods (capture, identify, etc.)
to `posthog/rails.rb` so they're available immediately when the gem
is required - before any initializers run.
