+++ posthog-rails/lib/posthog/rails/railtie.rb
@@ -3,76 +3,6 @@
 module PostHog
   module Rails
     class Railtie < ::Rails::Railtie
-      # Add PostHog module methods for accessing Rails-specific client
-      initializer 'posthog.set_configs' do |_app|
-        PostHog.class_eval do
-          class << self
-            attr_accessor :client
-
-            # Methods explicitly delegated to the client
-            DELEGATED_METHODS = %i[
-              capture
-              capture_exception
-              identify
-              alias
-              group_identify
-              is_feature_enabled
-              get_feature_flag
-              get_all_flags
-            ].freeze
-
-            # Initialize PostHog client with a block configuration
-            def init(options = {})
-              # If block given, yield to configuration
-              if block_given?
-                config = PostHog::Rails::InitConfig.new(options)
-                yield config
-                options = config.to_client_options
-              end
-
-              # Create the PostHog client
-              @client = PostHog::Client.new(options)
-            end
-
-            # Define delegated methods using metaprogramming
-            DELEGATED_METHODS.each do |method_name|
-              define_method(method_name) do |*args, **kwargs, &block|
-                ensure_initialized!
-                client.public_send(method_name, *args, **kwargs, &block)
-              end
-            end
-
-            def initialized?
-              !@client.nil?
-            end
-
-            # Fallback for any client methods not explicitly defined
-            # rubocop:disable Lint/RedundantSafeNavigation
-            def method_missing(method_name, ...)
-              if client&.respond_to?(method_name)
-                ensure_initialized!
-                client.public_send(method_name, ...)
-              else
-                super
-              end
-            end
-
-            def respond_to_missing?(method_name, include_private = false)
-              client&.respond_to?(method_name) || super
-            end
-            # rubocop:enable Lint/RedundantSafeNavigation
-
-            private
-
-            def ensure_initialized!
-              return if initialized?
-
-              raise 'PostHog is not initialized. Call PostHog.init in an initializer.'
-            end
-          end
-        end
-      end
-
       # Insert middleware for exception capturing
       initializer 'posthog.insert_middlewares' do |app|
         # Insert after DebugExceptions to catch rescued exceptions
